Se agregan tests

+++ src/triangulo/App.java
@@ -7,10 +7,10 @@ import java.util.Scanner;
 public class App {
 	public static void main(String[] args) {
 		try (Scanner lecturaPorTeclado = new Scanner(System.in)) {
-			List<Double> lados = new LinkedList<Double>();
+			List<Integer> lados = new LinkedList<Integer>();
 			for (int i = 0; i < 3; i++) {
 				System.out.println("Ingrese lado " + (i + 1) + ": ");
-				lados.add(lecturaPorTeclado.nextDouble());
+				lados.add(lecturaPorTeclado.nextInt());
 			}
 			Triangulo triangulo1 = new Triangulo(lados.get(0), lados.get(1), lados.get(2));
